moving files to a better file structure
