+++ 01_inventory/lib/inventory.rb
