+++ 01_inventory/store/in_file_store/data.yml
