+++ 01_inventory/store/in_memory_store.rb
